cleanup, public-ui test app.spec.ts to include all views and build the app.module
