+++ public-ui/src/app/views/app/app.component.html
@@ -7,7 +7,6 @@
     <div *ngIf="overriddenUrl" style="position: absolute; top: 0; left: 1rem;">
       <span style="font-size: 80%; color: darkred;">API URL: {{this.overriddenUrl}}</span>
     </div>
-    <app-routing-spinner></app-routing-spinner>
   </div>
 </clr-main-container>
